requirements.txt
openai==1.55.3
selenium==4.33.0
html5lib==1.1
python-dateutil==2.8.2
webdriver-manager==4.0.2
python-dotenv==1.1.0